Extract Serializer/Deserializer to separate files
