+++ modules/core/src/main/scala/vulcan/internal/Serializer.scala
@@ -1,0 +1,36 @@
+package vulcan
+package internal
+
+import org.apache.avro.generic.GenericDatumWriter
+import org.apache.avro.io.EncoderFactory
+
+import java.io.ByteArrayOutputStream
+import java.nio.charset.StandardCharsets
+
+private[vulcan] object Serializer {
+  final def toBinary[A](a: A)(implicit codec: Codec[A]): Either[AvroError, Array[Byte]] =
+    codec.schema.flatMap { schema =>
+      codec.encode(a).flatMap { encoded =>
+        AvroError.catchNonFatal {
+          val baos = new ByteArrayOutputStream
+          val encoder = EncoderFactory.get.binaryEncoder(baos, null)
+          new GenericDatumWriter[Any](schema).write(encoded, encoder)
+          encoder.flush()
+          Right(baos.toByteArray)
+        }
+      }
+    }
+
+  def toJson[A](a: A)(implicit codec: Codec[A]): Either[AvroError, String] =
+    codec.schema.flatMap { schema =>
+      codec.encode(a).flatMap { encoded =>
+        AvroError.catchNonFatal {
+          val baos = new ByteArrayOutputStream
+          val encoder = EncoderFactory.get.jsonEncoder(schema, baos)
+          new GenericDatumWriter[Any](schema).write(encoded, encoder)
+          encoder.flush()
+          Right(new String(baos.toByteArray, StandardCharsets.UTF_8))
+        }
+      }
+    }
+}
